Changed return type of getline to ssize_t
Also removed include of stdint.h since it is no longer
necessary.

+++ C/include/aoc_string.h
@@ -2,10 +2,9 @@
 #define __AOC_STRING_H__
 
 #include <stdio.h>
-#include <stdlib.h>
-#include <stdint.h>
 
 // String functions similar to methods of  python strings
+// more to come later
 char *str_trim(char *str);
 char *str_join(const char *, char **, size_t);
 int str_count(char *, char, int, int);
@@ -14,8 +13,8 @@ int str_startswith(char *str, char *start_str);
 int str_endswith(char *, char *);
 char *strdup_printf(const char *format, ...);
 
-intmax_t getline(char **, size_t *, FILE *);
-intmax_t getdelim(char **, size_t *, int, FILE *);
+ssize_t getline(char **, size_t *, FILE *);
+ssize_t getdelim(char **, size_t *, int, FILE *);
 char *stpcpy(char *__restrict__, const char *__restrict__);
 char *strdup(const char *str);
 char *strconcat(const char *str1, const char *str2);
